Add seasons page (#2)
Add a few handlers for viewing seasons. That involved:

* Initializing a DB connection pool on startup and adding it to the
`app` state
    * Still TODO: make the connection pool more configurable
* Adding a method to `app` to check out a DB connection from the pool
* Unifying the error types using
[`anyhow`](https://github.com/dtolnay/anyhow)
* Still TODO: use `anyhow`'s introspection features to differentiate
between different classes of errors when it comes to converting errors
into responses
* Establish a norm for how to structure related templates and handlers
* Fix the seasons schema to make several columns not nullable
* Settle on a serialization strategy for models to be used in views
(make the models themselves `impl Serialize`—in most cases this can be
done automatically)
    * Still TODO:
        * Add serializers for `chronological` types
* Make as many models as possible `derive(Serialize)`, and `impl
Serialize` for the ones that can't

cc @c1tadel I'm merging this directly for now since it sets up a lot of
structure that future features will want. Tomorrow I'm going to block
out a bunch of independent tasks and make a project board. Do you want
to chat this weekend about norms for collaborating on this/code
review/etc?